Move option-value component to dedicated folder
