+++ packages/harpguru-core/src/components/option-value/index.ts
@@ -1,0 +1,1 @@
+export { OptionValue } from './option-value'
